In-memory Engine: Adjust memory settings based on available memory
When the in-memory engine is enabled, it allocates 15% of the block
cache memory capacity as the `hard_limit_threshold`. If the
`soft_limit_threshold` is not explicitly configured, it defaults to 90%
of the `hard_limit_threshold`. Similarly, the `stop_load_limit_threshold`
defaults to 80% of the `hard_limit_threshold` if not manually set.

This adjustment optimizes memory usage based on the available resources.

Signed-off-by: Neil Shen <overvenus@gmail.com>